Refactor timezones endpoint to system.go
