+++ api4/system_test.go
@@ -527,3 +527,26 @@ func TestS3TestConnection(t *testing.T) {
 		t.Fatal("should return error ")
 	}
 }
+
+func TestSupportedTimezones(t *testing.T) {
+	th := Setup().InitBasic()
+	defer th.TearDown()
+	Client := th.Client
+
+	supportedTimezonesFromConfig := th.App.Config().SupportedTimezones
+	supportedTimezones, resp := Client.GetSupportedTimezone()
+
+	CheckNoError(t, resp)
+	for _, timezone := range supportedTimezones {
+		found := false
+		for _, configTimezone := range supportedTimezonesFromConfig {
+			if timezone == configTimezone {
+				found = true
+				break
+			}
+		}
+		if !found {
+			t.Fatalf("failed to find timezone: %v", timezone)
+		}
+	}
+}
