--- api4/timezone_test.go
@@ -1,29 +1,0 @@
-// Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
-
-package api4
-
-import "testing"
-
-func TestSupportedTimezones(t *testing.T) {
-	th := Setup().InitBasic()
-	defer th.TearDown()
-	Client := th.Client
-
-	supportedTimezonesFromConfig := th.App.Config().SupportedTimezones
-	supportedTimezones, resp := Client.GetSupportedTimezone()
-
-	CheckNoError(t, resp)
-	for _, timezone := range supportedTimezones {
-		found := false
-		for _, configTimezone := range supportedTimezonesFromConfig {
-			if timezone == configTimezone {
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Fatalf("failed to find timezone: %v", timezone)
-		}
-	}
-}
